requirements.txt
# Django Core
Django==5.1.3

# postgres adapter
psycopg[binary]==3.2.3

# Django Extensions
django-cors-headers==4.6.0 

# Django Rest Framework
djangorestframework==3.15.2
drf-spectacular==0.27.2

# flare library
py-flare-common==0.1.3

# web3
web3==7.5.0

# Sentry Monitoring
sentry-sdk==2.18.0

# Prometheus Monitoring
prometheus-client==0.21.0
django-prometheus==2.3.1

# dataclasses and serialization
attrs==24.2.0
cattrs==24.1.2

# wsgi
gunicorn==23.0.0
